+++ src/App.jsx
@@ -20,7 +20,7 @@ const App = () => {
           {/* 홈페이지는 <Route index element={<페이지 컴포넌트 />} /> 로 추가하기 */}
           {/* 각 페이지 <Route path="url" element={<페이지 컴포넌트 />} /> 로 추가하기 */}
           <Route path=":id/focus" element={<TodayFocus />} />
-          <Route path=":id/logs" element={<LogPage />} />
+          <Route path="/:studyId/logs" element={<LogPage />} />
           <Route path=":id/detail" element={<StudyDetailPage />} />
           <Route path=":id/habit" element={<TodayHabitPage />} />
           <Route path="/create" element={<StudyCreate />} />
