+++ src/pages/LogPage/LogPage.module.css
@@ -145,7 +145,7 @@
 }
 
 .rowValue {
-  color: var(--black_414141);
+  color: var(--green_text_578246);
   font-size: 16px;
 }
